Factor UnitVector out of ray.h into own header. Add some Ray closestApproach methods.
