+++ cpp/sophus/geometry/CMakeLists.txt
@@ -14,6 +14,7 @@ farm_ng_add_library(sophus_geometry
   plane_conv.h
   point_transform.h
   projection.h
+  unit_vector.h
   ray.h
 )
 target_link_libraries(sophus_geometry INTERFACE sophus_lie)
@@ -25,6 +26,7 @@ if(${BUILD_SOPHUS_TESTS})
     plane_conv
     point_transform
     projection
+    unit_vector
     ray)
     farm_ng_add_test(${test_basename}
                           PARENT_LIBRARY sophus_geometry
