+++ cpp/sophus/geometry/ray.h
@@ -9,8 +9,11 @@
 #pragma once
 
 #include "sophus/common/common.h"
+#include "sophus/geometry/unit_vector.h"
 #include "sophus/lie/se2.h"
 #include "sophus/lie/se3.h"
+#include "sophus/lie/sim2.h"
+#include "sophus/lie/sim3.h"
 
 #include <Eigen/Dense>
 #include <farm_ng/core/logging/expected.h>
@@ -22,19 +25,9 @@ namespace sophus {
 
 // Forward declarations
 template <class TScalar, int kN>
-class UnitVector;
-template <class TScalar, int kN>
 class Ray;
 
 // Convenience typedefs
-template <class TScalar>
-using UnitVector3 = UnitVector<TScalar, 3>;
-template <class TScalar>
-using UnitVector2 = UnitVector<TScalar, 2>;
-
-using UnitVector2F64 = UnitVector2<double>;
-using UnitVector3F64 = UnitVector3<double>;
-
 template <class TScalar>
 using Ray2 = Ray<TScalar, 2>;
 template <class TScalar>
@@ -42,50 +35,6 @@ using Ray3 = Ray<TScalar, 3>;
 
 using Ray2F64 = Ray2<double>;
 using Ray3F64 = Ray3<double>;
-
-template <class TScalar, int kN>
-class UnitVector {
- public:
-  // Precondition: v must be of unit length.
-  static UnitVector fromUnitVector(Eigen::Matrix<TScalar, kN, 1> const& v) {
-    farm_ng::Expected<UnitVector> e_vec = tryFromUnitVector(v);
-    if (!e_vec.has_value()) {
-      FARM_FATAL("{}", e_vec.error());
-    }
-    return e_vec.value();
-  }
-
-  static farm_ng::Expected<UnitVector> tryFromUnitVector(
-      Eigen::Matrix<TScalar, kN, 1> const& v) {
-    using std::abs;
-    if (abs(v.squaredNorm() - TScalar(1.0)) > sophus::kEpsilon<TScalar>) {
-      return FARM_ERROR(
-          "v must be unit length is {}.\n{}", v.squaredNorm(), v.transpose());
-    }
-    UnitVector unit_vector;
-    unit_vector.vector_ = v;
-    return unit_vector;
-  }
-
-  static UnitVector fromVectorAndNormalize(
-      Eigen::Matrix<TScalar, kN, 1> const& v) {
-    return fromUnitVector(v.normalized());
-  }
-
-  [[nodiscard]] Eigen::Matrix<TScalar, kN, 1> const& vector() const {
-    return vector_;
-  }
-
-  // Class invariant established from sibling we're copying from
-  UnitVector(UnitVector const&) = default;
-  UnitVector& operator=(UnitVector const&) = default;
-
- private:
-  UnitVector() {}
-
-  // Class invariant: v_ is of unit length.
-  Eigen::Matrix<TScalar, kN, 1> vector_;
-};
 
 template <class TScalar, int kN>
 class Ray {
@@ -138,20 +87,6 @@ class Ray {
 };
 
 template <class TScalar>
-inline UnitVector<TScalar, 2> operator*(
-    So2<TScalar> const& bar_rotation_foo, UnitVector<TScalar, 2> const& v_foo) {
-  return UnitVector<TScalar, 2>::fromUnitVector(
-      bar_rotation_foo * v_foo.vector());
-}
-
-template <class TScalar>
-inline UnitVector<TScalar, 3> operator*(
-    So3<TScalar> const& bar_rotation_foo, UnitVector<TScalar, 3> const& v_foo) {
-  return UnitVector<TScalar, 3>::fromUnitVector(
-      bar_rotation_foo * v_foo.vector());
-}
-
-template <class TScalar>
 inline Ray<TScalar, 2> operator*(
     Se2<TScalar> const& bar_pose_foo, Ray<TScalar, 2> const& ray_foo) {
   return Ray<TScalar, 2>(
@@ -167,6 +102,8 @@ inline Ray<TScalar, 3> operator*(
       bar_pose_foo.so3() * ray_foo.direction());
 }
 
+// Arbitrary 6-DoF transformation of a unit vector promotes it to a ray
+// having a potentially non-zero origin.
 template <class TScalar>
 inline Ray<TScalar, 3> operator*(
     Se3<TScalar> const& bar_pose_foo, UnitVector<TScalar, 3> const& v_foo) {
@@ -174,50 +111,54 @@ inline Ray<TScalar, 3> operator*(
       bar_pose_foo.translation(), bar_pose_foo.so3() * v_foo);
 }
 
-/// Construct rotation which would take unit direction vector ``from`` into
-/// ``to`` such that ``to = rotThroughPoints(from,to) * from``. I.e. that the
-/// rotated point ``from`` is colinear with ``to`` (equal up to scale)
-///
-/// The axis of rotation is perpendicular to both ``from`` and ``to``.
-///
-template <class TScalar>
-SOPHUS_FUNC So3<TScalar> rotThroughPoints(
-    UnitVector3<TScalar> const& from, UnitVector3<TScalar> const& to) {
-  using std::abs;
-  using std::atan2;
-  Eigen::Vector<TScalar, 3> from_cross_to = from.vector().cross(to.vector());
-  TScalar n = from_cross_to.norm();
-  if (abs(n) < sophus::kEpsilon<TScalar>) {
-    return So3<TScalar>();
-  }
-  // https://stackoverflow.com/a/32724066
-  TScalar angle = atan2(n, from.vector().dot(to.vector()));
-
-  return So3<TScalar>::exp(angle * from_cross_to / n);
+template <typename T>
+Ray2<T> operator*(Sim2<T> const& b_from_a, Ray2<T> const& ray_a) {
+  return Ray2<T>(
+      b_from_a * ray_a.origin(), b_from_a.rxso2() * ray_a.direction());
 }
 
-/// Construct rotation which would take direction vector ``from`` into ``to``
-/// such that ``to \propto rotThroughPoints(from,to) * from``. I.e. that the
-/// rotated point ``from`` is colinear with ``to`` (equal up to scale)
-///
-/// The axis of rotation is perpendicular to both ``from`` and ``to``.
-///
-/// Precondition: Neither ``from`` nor ``to`` must be zero. This is unchecked.
-template <class TScalar>
-SOPHUS_FUNC So3<TScalar> rotThroughPoints(
-    Eigen::Vector<TScalar, 3> const& from,
-    Eigen::Vector<TScalar, 3> const& to) {
-  using std::abs;
-  using std::atan2;
-  Eigen::Vector<TScalar, 3> from_cross_to = from.cross(to);
-  TScalar n = from_cross_to.norm();
-  if (abs(n) < sophus::kEpsilon<TScalar>) {
-    return So3<TScalar>();
-  }
-  // https://stackoverflow.com/a/32724066
-  TScalar angle = atan2(n, from.dot(to));
+template <typename T>
+Ray3<T> operator*(Sim3<T> const& b_from_a, Ray3<T> const& ray_a) {
+  return Ray3<T>(
+      b_from_a * ray_a.origin(), b_from_a.rxso3() * ray_a.direction());
+}
 
-  return So3<TScalar>::exp(angle * from_cross_to / n);
+// For two lines:
+//   line_a: x = A + lambda * B
+//   line_b: y = C + mu * D
+// returns distances [lambda, mu] along the respective rays, corresponding to
+// the closest approach of x and y according to an l2 distance measure. lambda
+// and mu may be positive or negative.
+//
+// TODO: what if they are parallel? Presumably E * line_a.direction() will be 0.
+//       should we return an optional?
+template <class T>
+std::array<T, 2> closestApproachDistances(
+    Ray3<T> const& line_a, Ray3<T> const& line_b) {
+  // E = B.D.D^T \in R^3
+  const Eigen::Matrix<T, 1, 3> E = line_a.direction().transpose() *
+                                   line_b.direction() *
+                                   line_b.direction().transpose();
+
+  // lambda = E*(A-C) / (E*B)
+  const T lambda =
+      E * (line_a.origin() - line_b.origin()) / (E * line_a.direction());
+
+  // mu = (A-C-lambda*B)*D
+  const T mu =
+      (line_a.origin() - line_b.origin() - lambda * line_a.direction()) *
+      line_b.direction();
+  return {lambda, mu};
+}
+
+// For two lines ``line_a`` and ``line_b`` returns the mid-point of the line
+// segment connecting one point from each of the lines which are closest to
+// one another according to the l2 distance measure.
+template <class T>
+Eigen::Vector3<T> closestApproach(
+    Ray3<T> const& line_a, Ray3<T> const& line_b) {
+  auto [lambda, mu] = closestApproachDistances(line_a, line_b);
+  return (line_a.pointAt(lambda) + line_b.pointAt(mu)) / static_cast<T>(2.0);
 }
 
 }  // namespace sophus
